# bilemo_api_project/public/swagger/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Api BileMo",
        "version": "0.1"
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "Api BileMo"
        }
    ],
    "paths": {
        "/customers/{username}": {
            "get": {
                "tags": [
                    "Customers"
                ],
                "operationId": "App\\Controller\\CustomerController::show",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/username"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Show a customer ressource",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Customer"
                                }
                            }
                        }
                    },
                    "403": {
                        "$ref": "#/components/responses/Unauthorized"
                    }
                },
                "security": [
                    {
                        "scalar": "bearer"
                    }
                ]
            }
        },
        "/customers": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "operationId": "App\\Controller\\CustomerController::index",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/page"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of customers ressources",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Customer"
                                    }
                                }
                            }
                        }
                    },
                    "403": {
                        "$ref": "#/components/responses/Unauthorized"
                    }
                },
                "security": [
                    {
                        "scalar": "bearer"
                    }
                ]
            }
        },
        "/customers/{username]/users": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "operationId": "App\\Controller\\CustomerController::customerUserList",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/username"
                    },
                    {
                        "$ref": "#/components/parameters/page"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of users ressource linked to a customer",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/User"
                                    }
                                }
                            }
                        }
                    },
                    "403": {
                        "$ref": "#/components/responses/Unauthorized"
                    }
                },
                "security": [
                    {
                        "scalar": "bearer"
                    }
                ]
            }
        },
        "/phones/{id}": {
            "get": {
                "tags": [
                    "Phones"
                ],
                "operationId": "App\\Controller\\PhoneController::show",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/id"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Shows a Phone",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Phone"
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    }
                },
                "security": [
                    {
                        "scalar": "bearer"
                    }
                ]
            },
            "put": {
                "tags": [
                    "Phones"
                ],
                "operationId": "App\\Controller\\PhoneController::update",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/id"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Phone Update",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Phone"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "$ref": "#/components/responses/BadRequest"
                    }
                },
                "security": [
                    {
                        "scalar": "bearer"
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Phones"
                ],
                "operationId": "App\\Controller\\PhoneController::delete",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/id"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Delete a Phone"
                    },
                    "403": {
                        "$ref": "#/components/responses/Unauthorized"
                    }
                },
                "security": [
                    {
                        "scalar": "bearer"
                    }
                ]
            }
        },
        "/phones": {
            "get": {
                "tags": [
                    "Phones"
                ],
                "operationId": "App\\Controller\\PhoneController::index",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/page"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of phones",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Phone"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "scalar": "bearer"
                    }
                ]
            },
            "post": {
                "tags": [
                    "Phones"
                ],
                "responses": {
                    "201": {
                        "description": "New phone ressource created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "New phone ressource created"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "$ref": "#/components/responses/BadRequest"
                    }
                },
                "security": [
                    {
                        "scalar": "bearer"
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Customer": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "username": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "users": {
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "Phone": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "price": {
                        "type": "integer"
                    },
                    "color": {
                        "type": "string"
                    },
                    "releasedAt": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "User": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "firstName": {
                        "type": "string"
                    },
                    "lastName": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "customer": {
                        "type": "object"
                    }
                },
                "type": "object"
            }
        },
        "responses": {
            "NotFound": {
                "description": "Ressource not found",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "example": "user does not exists"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "Unauthorized": {
                "description": "Access denied",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "example": "You do not have the permissions to access this endpoint"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "BadRequest": {
                "description": "Bad Request",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "example": "A certain field was missing in the body"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            }
        },
        "parameters": {
            "id": {
                "name": "id",
                "in": "path",
                "description": "Ressource id property",
                "required": true,
                "schema": {
                    "type": "integer"
                }
            },
            "username": {
                "name": "username",
                "in": "path",
                "description": "Ressource username property",
                "required": true,
                "schema": {
                    "type": "string"
                }
            },
            "page": {
                "name": "page",
                "in": "query",
                "description": "the current page",
                "required": false,
                "schema": {
                    "type": "integer"
                }
            }
        },
        "securitySchemes": {
            "bearer": {
                "type": "apiKey",
                "name": "BEARER",
                "in": "cookie",
                "bearerFormat": "JWT"
            }
        }
    }
}